requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
joblib==1.5.1
nltk==3.9.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
textblob==0.19.0
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
vaderSentiment==3.3.2
